fix: move dir to kart
