+++ src/kart/UserMatchList.ts
@@ -1,20 +1,19 @@
 
 import UserMatchInfo from './UserMatchInfo'; 
 import type { UserBasicInfo } from './KartClient'
-import type { rawMatchTypeMatches, rawUserMatchInfo, rawUserMatchList } from '../typings/raw'
+import type { rawMatchTypeMatches, rawUserMatchInfo, rawUserMatchList } from '../../typings/raw'
 
 export default class UserMatchList{
     public user: UserBasicInfo;
     public matchList: UserMatchInfo[];
-    public limit: number | undefined;
-    public offset: number | undefined;
+    public limit: number | undefined | null;
+    public offset: number | undefined | null;
     constructor(userInfo:UserBasicInfo, data: rawUserMatchList, page?:{limit:number,offset:number}){
         this.user = userInfo;
         this.matchList = resolveMatchList(data.matches);
         this.limit = page?.limit;
         this.offset = page?.offset;
-    }
-    
+    }    
 }
 
 function _MargeTypefromMatchObject(data: rawMatchTypeMatches[]){
