+++ test/getUserMatchList.ts
@@ -5,6 +5,8 @@ import _ from 'lodash';
 const client = new KartClient(token);
 (async () =>{
 const userinfo = await client.getUserBasicInfoByName('코로나행열차');
+if(!userinfo){console.log('userinfo is null'); return;}
 const matchlist = await client.getUserMatchList(userinfo);
-console.log(matchlist)
+if(!matchlist) return;
+console.log(matchlist.matchList[0])
 })();
